+++ backend/src/main/java/middle_point_search/backend/common/security/dto/CustomUserDetails.java
@@ -1,64 +1,7 @@
 package middle_point_search.backend.common.security.dto;
 
-import java.util.ArrayList;
-import java.util.Collection;
-
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class CustomUserDetails implements UserDetails {
-
-	@Getter
-	private String placeRoomId;
-	private String name;
-	private String pw;
-	private String authority;
-	private boolean enabled;
-
-	public static CustomUserDetails of(String placeRoomId, String name, String pw, String authority, boolean enabled) {
-		return new CustomUserDetails(placeRoomId, name, pw, authority, enabled);
-	}
-
-	@Override
-	public Collection<? extends GrantedAuthority> getAuthorities() {
-		ArrayList<GrantedAuthority> auth = new ArrayList<>();
-		auth.add(new SimpleGrantedAuthority(authority));
-		return auth;
-	}
-
-	@Override
-	public String getPassword() {
-		return pw;
-	}
-
-	@Override
-	public String getUsername() {
-		return name;
-	}
-
-	@Override
-	public boolean isAccountNonExpired() {
-		return true;
-	}
-
-	@Override
-	public boolean isAccountNonLocked() {
-		return true;
-	}
-
-	@Override
-	public boolean isCredentialsNonExpired() {
-		return true;
-	}
-
-	@Override
-	public boolean isEnabled() {
-		return enabled;
-	}
+public interface CustomUserDetails extends UserDetails {
+	String getRoomId();
 }
